src-licensing/pyproject.toml
[project]
name = "promptshield-licensing"
version = "1.0.0"
description = "Licensing & billing server for promptShield"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "stripe>=8.0",
    "pynacl>=1.5",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]
